testing

+++ Lab5/code/bookstore/Book.java
@@ -1,19 +1,20 @@
 //package bookstore;
+import java.util.*;
 
 public class Book {
     private String title;
     private String author;
     private Date publicationDate;
     private String publicationPlace;
-    private int isbn;
+    private long isbn;
 
-    public Book( String ti, String ai, Date pdi, String ppi, int isbni ) {
+    public Book( String ti, String ai, Date pdi, String ppi, long isbni ) {
         title = ti; author = ai; publicationDate = pdi; publicationPlace = ppi; isbn = isbni;
     }
     public String getTitle() { return title; }
     public String getAuthor() { return author; }
     public Date getPublicationDate() { return publicationDate; }
     public String getPublicationPlace() { return publicationPlace; }
-    public int getISBN() { return isbn; }
+    public long getISBN() { return isbn; }
 
 }
